refactor: merge Diacritics with TextUtils
