--- RpFilter/Diacritics.lua
@@ -1,10 +1,0 @@
--- Standard word characters extended with diacritics
-WORD_CHARS = "A-Za-z0-9" ..
-            "ÀÁÂÃÄÅÆÇÈÉÊËÌÍÎÏÐÑÒÓÔÕÖØÙÚÛÜÝÞß" ..
-            "àáâãäåæçèéêëìíîïðñòóôõöøùúûüýþÿ"
-
----@param word string
----@return boolean
-function IsCapitalized(word)
-    return word:match("^'?[A-ZÀÁÂÃÄÅÆÇÈÉÊËÌÍÎÏÐÑÒÓÔÕÖØÙÚÛÜÝÞß]") ~= nil
-end
